requirements.txt
guppy3==3.1.0
memory-profiler==0.58.0
numpy==1.20.1
opencv-python==4.5.1.48
pandas==1.2.3
Pillow==8.1.1
psutil==5.8.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
typing-extensions==3.7.4.3
